# swagger/openapi.yaml
openapi: 3.1.0
info:
  title: Immersive Images
  version: 1.0.0
tags:
  - name: images
  - name: system
paths:
  /images:
    post:
      tags: [images]
      operationId: createImage
      summary: Create an image
      description: |
        Generates an image from the given seed, uploads it to Selectel S3 and
        returns the stored metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'components.yaml#/components/schemas/CreateImageRequest'
      responses:
        '201':
          description: Image created successfully.
          content:
            application/json:
              schema:
                $ref: 'components.yaml#/components/schemas/ImageMeta'
        '400':
          $ref: 'components.yaml#/components/responses/BadRequest'
        '500':
          $ref: 'components.yaml#/components/responses/InternalServerError'
        '503':
          $ref: 'components.yaml#/components/responses/Unavailable'
    get:
      tags: [images]
      operationId: listImages
      summary: List images
      description: Returns a paginated list of images, newest first.
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
        - name: offset
          in: query
          description: Number of items to skip.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated list of images.
          content:
            application/json:
              schema:
                $ref: 'components.yaml#/components/schemas/ListImagesResponse'
        '400':
          $ref: 'components.yaml#/components/responses/BadRequest'
        '500':
          $ref: 'components.yaml#/components/responses/InternalServerError'
  /images/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        description: UUID of the image.
        schema:
          type: string
          format: uuid
    get:
      tags: [images]
      operationId: getImage
      summary: Get image metadata
      description: Returns metadata of a single image.
      responses:
        '200':
          description: Image metadata.
          content:
            application/json:
              schema:
                $ref: 'components.yaml#/components/schemas/ImageMeta'
        '404':
          $ref: 'components.yaml#/components/responses/NotFound'
        '500':
          $ref: 'components.yaml#/components/responses/InternalServerError'
    put:
      tags: [images]
      operationId: updateImage
      summary: Rename an image
      description: Updates the name of the image.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'components.yaml#/components/schemas/UpdateImageRequest'
      responses:
        '200':
          description: Updated image metadata.
          content:
            application/json:
              schema:
                $ref: 'components.yaml#/components/schemas/ImageMeta'
        '400':
          $ref: 'components.yaml#/components/responses/BadRequest'
        '404':
          $ref: 'components.yaml#/components/responses/NotFound'
        '500':
          $ref: 'components.yaml#/components/responses/InternalServerError'
        '503':
          $ref: 'components.yaml#/components/responses/Unavailable'
    delete:
      tags: [images]
      operationId: deleteImage
      summary: Delete an image
      description: Deletes the image from storage.
      responses:
        '204':
          description: Image deleted successfully.
        '404':
          $ref: 'components.yaml#/components/responses/NotFound'
        '500':
          $ref: 'components.yaml#/components/responses/InternalServerError'
        '503':
          $ref: 'components.yaml#/components/responses/Unavailable'
  /images/{uuid}/content:
    parameters:
      - name: uuid
        in: path
        required: true
        description: UUID of the image.
        schema:
          type: string
          format: uuid
    get:
      tags: [images]
      operationId: getImageContent
      summary: Download image file
      description: |
        Downloads the image file as PNG. Image content is immutable — it is
        generated once from the seed and never changes afterwards.
      responses:
        '200':
          description: The image file.
          headers:
            Cache-Control:
              description: |
                Content never changes, so the file may be cached indefinitely
                without revalidation.
              schema:
                type: string
                example: 'public, max-age=31536000, immutable'
            Content-Disposition:
              description: Suggested download filename.
              schema:
                type: string
                example: 'attachment; filename="my-image.png"'
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          $ref: 'components.yaml#/components/responses/NotFound'
        '500':
          $ref: 'components.yaml#/components/responses/InternalServerError'
  /health:
    get:
      tags: [system]
      operationId: health
      summary: Health check
      description: Returns service health status.
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok